user routes updated unauthorized

+++ server/api/users.js
@@ -28,7 +28,12 @@ router.get('/', checkIfAdmin, async (req, res, next) => {
 router.get('/:id', async (req, res, next) => {
   try {
     const thisUser = await User.findByPk(req.params.id)
-    if (thisUser && req.user.email === thisUser.email) res.json(thisUser)
+    if (!req.user) res.sendStatus(401)
+    if (
+      (thisUser && req.user.email === thisUser.email) ||
+      (req.user && req.user.accountType === 'Admin')
+    )
+      res.json(thisUser)
     else res.sendStatus(401)
   } catch (err) {
     next(err)
@@ -47,11 +52,18 @@ router.post('/', checkIfAdmin, async (req, res, next) => {
 
 router.delete('/:id', checkIfAdmin, async (req, res, next) => {
   try {
-    await User.destroy({
-      where: {
-        id: req.params.id
-      }
-    })
+    const thisUser = await User.findByPk(req.params.id)
+    if (!req.user) res.sendStatus(401)
+    if (
+      (thisUser && req.user.email === thisUser.email) ||
+      (req.user && req.user.accountType === 'Admin')
+    ) {
+      await User.destroy({
+        where: {
+          id: req.params.id
+        }
+      })
+    }
   } catch (err) {
     next(err)
   }
@@ -60,7 +72,11 @@ router.delete('/:id', checkIfAdmin, async (req, res, next) => {
 router.put('/:id', async (req, res, next) => {
   try {
     const thisUser = await User.findByPk(req.params.id)
-    if (thisUser && req.user.email === thisUser.email) {
+    if (!req.user) res.sendStatus(401)
+    if (
+      (thisUser && req.user.email === thisUser.email) ||
+      (req.user && req.user.accountType === 'Admin')
+    ) {
       await thisUser.update(req.body)
       res.json(thisUser)
     } else res.sendStatus(401)
